+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add minimum channel size
 - Split onboading and create invoice flow
 - Format amounts in input fields
+- Return to wallet dashboard on claimed payment
 
 ## [1.3.0] - 2023-09-27